--- sage-android/Sage/app/src/main/java/blueprint/com/sage/checkIn/fragments/CheckInListAbstractFragment.java
@@ -1,12 +1,0 @@
-package blueprint.com.sage.checkIn.fragments;
-
-import android.support.v4.app.Fragment;
-
-import blueprint.com.sage.checkIn.CheckInListActivity;
-
-/**
- * Created by charlesx on 11/10/15.
- */
-public class CheckInListAbstractFragment extends Fragment {
-    public CheckInListActivity getParentActivity() { return (CheckInListActivity) getActivity(); }
-}
