+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/requests/fragments/RequestsAbstractFragment.java
@@ -1,0 +1,12 @@
+package blueprint.com.sage.requests.fragments;
+
+import android.support.v4.app.Fragment;
+
+import blueprint.com.sage.requests.RequestsActivity;
+
+/**
+ * Created by charlesx on 11/10/15.
+ */
+public class RequestsAbstractFragment extends Fragment {
+    public RequestsActivity getParentActivity() { return (RequestsActivity) getActivity(); }
+}
