--- sage-android/Sage/app/src/main/java/blueprint/com/sage/shared/AbstractFragment.java
@@ -1,7 +1,0 @@
-package blueprint.com.sage.shared;
-
-/**
- * Created by charlesx on 10/12/15.
- */
-public class AbstractFragment {
-}
