+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/signUp/views/NoSwipeViewPager.java
@@ -19,5 +19,4 @@ public class NoSwipeViewPager extends ViewPager {
 
     @Override
     public boolean onTouchEvent(MotionEvent event) { return false; }
-
 }
